Added 'tutorial_5/strconcat.py'
Demonstrate the best way to create strings from a collection of elements.

The fastest way is to use the join() method on a list.

Avoid generators: a generator is converted into a list first.

This is because the join() method needs to walk on the list two times:
the first to get the actual size of the list;
the second to actual copy the data.